Update benchmark.h
Generating 400M read ops is impossible when you load only 400000 initial keys hence the seg faults. To ensure we run the operation until the index is full, we have adjusted the number of runs, in this case, the initial ops to be generated should be the starting number of keys already loaded multiplied by 2.This ensures keys are inserted at multiples of the initial ops/ 2 = initial keys loaded.